Added cscope related stuff
Added a script to prepare cscope files in any directory

It is only for .c and .h files but can be expanded to contain C++
related files too.

Added keys to navigate in vim. Also vim auto-loads the cscope file.
There is a better way which looks for a cscope file higher in fs
hierarchy which is pretty useful. I'll add that too someday.